+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support for New compute shape E4-Dense
 - Support for Cloud Advisor Fix It actions for V1 recommendations
+- Support for Cloud Advisor customization at resource/compartment level
 
 ## 4.14.0 (February 17, 2021)